requirements.txt
annotated-types==0.7.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy
cycler==0.12.1
dotenv==0.9.9
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud==0.34.0
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-cloud-vision==3.11.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
Jinja2==3.1.6
kiwisolver
MarkupSafe==3.0.3
matplotlib
mpmath==1.3.0
networkx==3.6
numpy
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pip==23.0.1
polars==1.35.2
polars-runtime-32==1.35.2
proto-plus==1.26.1
protobuf==5.29.5
psutil
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy
setuptools==66.1.1
simpleaudio
six==1.17.0
sounddevice==0.5.3
sympy==1.14.0
torch
torchaudio
torchvision
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0